Add shared HTTP client for integration tools
- Introduced `IntegrationClient` to manage backend interactions, including methods for POST and GET requests with error handling.
- Implemented caching for pricing information fetched from the backend.
- Created a new `types` module to define shared types such as `BackendResponse` and `IntegrationPricing`, enhancing code organization and clarity.
- Updated `mod.rs` to include the new client and types, ensuring proper module structure.

These changes improve the maintainability and functionality of the integration tools by providing a robust client for backend communication.
